+++ activesupport/lib/active_support/core_ext/kernel/reporting.rb
@@ -1,4 +1,5 @@
 require 'rbconfig'
+require 'tempfile'
 
 module Kernel
   # Sets $VERBOSE to nil for the duration of the block and back to its original
@@ -66,19 +67,33 @@ module Kernel
 
   # Captures the given stream and returns it:
   #
-  #   stream = capture(:stdout) { puts 'Cool' }
-  #   stream # => "Cool\n"
+  #   stream = capture(:stdout) { puts 'notice' }
+  #   stream # => "notice\n"
+  #
+  #   stream = capture(:stderr) { warn 'error' }
+  #   stream # => "error\n"
+  #
+  # even for subprocesses:
+  #
+  #   stream = capture(:stdout) { system('echo notice') }
+  #   stream # => "notice\n"
+  #
+  #   stream = capture(:stderr) { system('echo error 1>&2') }
+  #   stream # => "error\n"
   def capture(stream)
-    begin
-      stream = stream.to_s
-      eval "$#{stream} = StringIO.new"
-      yield
-      result = eval("$#{stream}").string
-    ensure
-      eval("$#{stream} = #{stream.upcase}")
-    end
+    stream = stream.to_s
+    captured_stream = Tempfile.new(stream)
+    stream_io = eval("$#{stream}")
+    origin_stream = stream_io.dup
+    stream_io.reopen(captured_stream)
 
-    result
+    yield
+
+    stream_io.rewind
+    return captured_stream.read
+  ensure
+    captured_stream.unlink
+    stream_io.reopen(origin_stream)
   end
   alias :silence :capture